fix(skills): give Supervisor lifecycle calls their own path, split the contract file
Codex round 13.

The lifecycle row sent hassio.addon_start|stop|restart and
host_reboot|shutdown into a Flow that resolves an entity, reads
/api/states, and verifies the same entity afterwards. An App is
addressed by slug, has no entity, and a host reboot has no target at all
— it takes the transport down, so nothing can be verified from here. All
five now have an explicit path: slug resolution from /addons, read-back
from /addons/<slug>/info, and for the host calls an honest "reported as
issued, never as succeeded, and shutdown needs physical access to come
back".

And the file was 558 lines against the repo's ~400 ceiling. The #513
gate contract moved to its own file — same subject, own surface — and
the seven constants that only the moved blocks used are gone. 235 and
323 lines now.
