+++ scripts/test/safe-core-files.json
@@ -91,6 +91,7 @@
   "tests/skills/health-source-shapes.test.ts",
   "tests/skills/helper-contract.test.ts",
   "tests/skills/home-status-redesign-contract.test.ts",
+  "tests/skills/indirect-actuation-contract.test.ts",
   "tests/skills/integration-setup-contract.test.ts",
   "tests/skills/maintenance-contract.test.ts",
   "tests/skills/output-design-contract.test.ts",
